+++ CHANGELOG.md
@@ -6,6 +6,8 @@ and this project adheres to
 
 ## [Unreleased]
 
+## [0.28.10] - 2022-06-29
+
 ### Fixed
 
 - @cosmjs/tendermint-rpc: Fix decoding events without attributes ([#1198]).
@@ -929,7 +931,8 @@ CHANGELOG entries missing. Please see [the diff][0.24.1].
   `FeeTable`. @cosmjs/cosmwasm has its own `FeeTable` with those properties.
 - @cosmjs/sdk38: Rename package to @cosmjs/launchpad.
 
-[unreleased]: https://github.com/cosmos/cosmjs/compare/v0.28.9...HEAD
+[unreleased]: https://github.com/cosmos/cosmjs/compare/v0.28.10...HEAD
+[0.28.10]: https://github.com/cosmos/cosmjs/compare/v0.28.9...v0.28.10
 [0.28.9]: https://github.com/cosmos/cosmjs/compare/v0.28.8...v0.28.9
 [0.28.8]: https://github.com/cosmos/cosmjs/compare/v0.28.7...v0.28.8
 [0.28.7]: https://github.com/cosmos/cosmjs/compare/v0.28.6...v0.28.7